Fix about page and path issues. Update paths to serve files from root instead of /public/, create about.html at root level for direct access, add partials to correct location, and fix linting issues.
