+++ src/about.html
@@ -1,0 +1,85 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>About Us - SleepOutside</title>
+    <link rel="stylesheet" href="./css/style.css">
+    <link rel="stylesheet" href="./css/about.css">
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css">
+    <script src="./js/about.js" type="module"></script>
+</head>
+<body>
+    <header>
+        <!-- Header will be loaded by JavaScript -->
+    </header>
+
+    <main class="about-page">
+        <section class="hero">
+            <div class="hero-content">
+                <h1>Our Story</h1>
+                <p>Providing quality outdoor gear for adventurers since 2010</p>
+            </div>
+        </section>
+
+        <section class="about-content">
+            <div class="container">
+                <div class="about-grid">
+                    <div class="about-text">
+                        <h2>Who We Are</h2>
+                        <p>SleepOutside was founded with a simple mission: to provide high-quality outdoor gear that helps people connect with nature. What started as a small shop in the mountains has grown into a trusted name in outdoor equipment.</p>
+                        
+                        <h2>Our Values</h2>
+                        <div class="values">
+                            <div class="value-item">
+                                <i class="fas fa-mountain"></i>
+                                <h3>Quality</h3>
+                                <p>We source only the best gear from trusted manufacturers.</p>
+                            </div>
+                            <div class="value-item">
+                                <i class="fas fa-leaf"></i>
+                                <h3>Sustainability</h3>
+                                <p>We're committed to protecting the outdoors we love.</p>
+                            </div>
+                            <div class="value-item">
+                                <i class="fas fa-users"></i>
+                                <h3>Community</h3>
+                                <p>We're proud to support local outdoor communities.</p>
+                            </div>
+                        </div>
+                    </div>
+                    
+                    <div class="about-image">
+                        <img src="./images/about-hero.jpg" alt="Outdoor adventure" class="about-hero-img">
+                    </div>
+                </div>
+
+                <div class="team-section">
+                    <h2>Meet Our Team</h2>
+                    <div class="team-grid">
+                        <div class="team-member">
+                            <img src="./images/team1.jpg" alt="Team Member 1">
+                            <h3>Alex Johnson</h3>
+                            <p>Founder & CEO</p>
+                        </div>
+                        <div class="team-member">
+                            <img src="./images/team2.jpg" alt="Team Member 2">
+                            <h3>Jamie Smith</h3>
+                            <p>Head of Product</p>
+                        </div>
+                        <div class="team-member">
+                            <img src="./images/team3.jpg" alt="Team Member 3">
+                            <h3>Taylor Brown</h3>
+                            <p>Customer Experience</p>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </section>
+    </main>
+
+    <footer>
+        <!-- Footer will be loaded by JavaScript -->
+    </footer>
+</body>
+</html>
