+++ src/cart/cart.js
@@ -1,6 +1,23 @@
 import { loadHeaderFooter } from "../js/utils.mjs";
 import ShoppingCart from "../js/ShoppingCart.mjs";
 
+// Function to update active navigation link
+function updateActiveNav() {
+  const currentPath = window.location.pathname;
+  const navLinks = document.querySelectorAll(".nav-link");
+  
+  navLinks.forEach(link => {
+    const linkPath = link.getAttribute("href");
+    // Check if the current path includes the link path or if we're on the home page
+    if ((currentPath === "/" && linkPath === "/index.html") || 
+        (currentPath !== "/" && linkPath !== "/index.html" && currentPath.includes(linkPath.replace("/index.html", "")))) {
+      link.classList.add("active");
+    } else {
+      link.classList.remove("active");
+    }
+  });
+}
+
 // Initialize the shopping cart
 const cart = ShoppingCart.getInstance();
 
@@ -8,6 +25,7 @@ const cart = ShoppingCart.getInstance();
 document.addEventListener("DOMContentLoaded", async () => {
   try {
     await loadHeaderFooter();
+    updateActiveNav();
     await cart.renderCart();
     
     // Add any additional cart page initialization here
