+++ src/css/header.css
@@ -8,6 +8,8 @@ header {
   box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
   position: relative;
   z-index: 100;
+  flex-wrap: wrap;
+  gap: 1rem;
 }
 
 .logo {
@@ -30,6 +32,49 @@ header {
 
 .highlight {
   color: #2c7be5;
+}
+
+/* Navigation Styles */
+.main-nav ul {
+  display: flex;
+  list-style: none;
+  margin: 0;
+  padding: 0;
+  gap: 1.5rem;
+}
+
+.main-nav a {
+  text-decoration: none;
+  color: #333;
+  font-weight: 500;
+  padding: 0.5rem 0;
+  position: relative;
+  transition: color 0.3s ease;
+}
+
+.main-nav a:hover {
+  color: #2c7be5;
+}
+
+.main-nav a::after {
+  content: '';
+  position: absolute;
+  width: 0;
+  height: 2px;
+  bottom: 0;
+  left: 0;
+  background-color: #2c7be5;
+  transition: width 0.3s ease;
+}
+
+.main-nav a:hover::after,
+.main-nav a.active::after {
+  width: 100%;
+}
+
+.main-nav a.active {
+  color: #2c7be5;
+  font-weight: 600;
 }
 
 /* Cart Styles */
@@ -61,34 +106,47 @@ header {
   position: absolute;
   top: -5px;
   right: -5px;
-  background-color: #e74c3c;
+  background-color: #2c7be5;
   color: white;
   border-radius: 50%;
   width: 20px;
   height: 20px;
   display: flex;
-  justify-content: center;
   align-items: center;
-  font-size: 0.75rem;
+  justify-content: center;
+  font-size: 0.7rem;
   font-weight: bold;
-  transform: scale(0.8);
-  transition: transform 0.2s ease;
 }
 
 .cart-link:hover .cart-count {
   transform: scale(1);
 }
 
-/* Responsive Adjustments */
+/* Responsive Styles */
 @media (max-width: 768px) {
   header {
-    padding: 0.75rem 1rem;
+    flex-direction: column;
+    padding: 1rem;
   }
-
+  
+  .main-nav ul {
+    margin: 1rem 0;
+    flex-wrap: wrap;
+    justify-content: center;
+  }
+  
+  .logo {
+    margin-bottom: 0.5rem;
+  }
+  
   .logo a {
     font-size: 1.25rem;
   }
-
+  
+  .cart {
+    margin-top: 0.5rem;
+  }
+  
   .cart-icon {
     width: 28px;
     height: 28px;
@@ -98,5 +156,16 @@ header {
     width: 18px;
     height: 18px;
     font-size: 0.65rem;
+  }
+}
+
+@media (max-width: 480px) {
+  .main-nav ul {
+    gap: 1rem;
+  }
+  
+  .main-nav a {
+    font-size: 0.9rem;
+    padding: 0.3rem 0;
   }
 }
