+++ src/index.html
@@ -8,7 +8,7 @@
   </head>
 
   <body>
-    <!--#include virtual="/public/partials/header.html" -->
+    <!--#include virtual="/partials/header.html" -->
     <main class="divider">
       <div class="hero">
         <img
@@ -57,7 +57,7 @@
         </ul>
       </section>
     </main>
-    <!--#include virtual="/public/partials/footer.html" -->
+    <!--#include virtual="/partials/footer.html" -->
     <script src="/js/main.js" type="module"></script>
   </body>
 </html>
