+++ src/js/utils.mjs
@@ -60,7 +60,7 @@ export async function loadHeaderFooter() {
   // Only load templates if the elements don't already contain content
   if (!headerElement || !headerElement.innerHTML.trim()) {
     try {
-      const header = await loadTemplate("/public/partials/header.html");
+      const header = await loadTemplate("/partials/header.html");
       if (headerElement) {
         renderWithTemplate(header, headerElement);
       }
@@ -71,7 +71,7 @@ export async function loadHeaderFooter() {
 
   if (!footerElement || !footerElement.innerHTML.trim()) {
     try {
-      const footer = await loadTemplate("/public/partials/footer.html");
+      const footer = await loadTemplate("/partials/footer.html");
       if (footerElement) {
         renderWithTemplate(footer, footerElement);
       }
