+++ src/partials/footer.html
@@ -1,0 +1,4 @@
+<footer id="main-footer">
+  &copy;2025 ⛺ SleepOutside ⛺ WDD 330 ⛺ BYU-Idaho for BYU-Pathway
+  Worldwide Online
+</footer>
