+++ src/product_listing/index.html
@@ -8,7 +8,7 @@
   </head>
 
   <body>
-    <!--#include virtual="/public/partials/header.html" -->
+    <!--#include virtual="/partials/header.html" -->
     <main class="divider">
       <section class="products">
         <h2>Our Products</h2>
@@ -17,7 +17,7 @@
         </div>
       </section>
     </main>
-    <!--#include virtual="/public/partials/footer.html" -->
+    <!--#include virtual="/partials/footer.html" -->
     <script src="/js/product-listing.js" type="module"></script>
   </body>
 </html>
